requirements.txt
# QuantumLeap 2026 ICCP - Responsible AI Connector
# Python 3.10+

# Web & API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.2

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# RAG & embeddings
pypdf>=4.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0

# LLM (optional: use OpenAI-compatible API or local)
httpx>=0.26.0
openai>=1.12.0

# FastAPI-MCP (server-side tools for LLM)
fastapi-mcp>=0.3.0

# Utils
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
